Update patch set 1
Patch Set 1:

> (1 comment)

There was a decision to make the directory under /dockerdata-nfs to be "Release.Name", which is the name of the helm deployed release and not the namespace. It will provide the ability to deploy on the same environment in the same namespace a "test" helm release and after testing "prod" helm release.

Patch-set: 1