Update patch set 1
Patch Set 1:

Jessica, Yes the patch is changing the reference to onap/ccsdk-dgbuilder-image:0.2.1
but the root cause is to understand what happened to:

onap/ccsdk-dgbuilder-image:0.2.1-SNAPSHOT
From a search right now, it is still there but the JIRA ticket was opened because the CD system found that 0.2.1-SNAPSHOT was unavailable for a period of time.

Since the original -SNAPSHOT image is back can we abandon this patch?  The image the HEAT version uses 0.2.1-SNAPSHOT so I think that is the correct tag.

Patch-set: 1
Reviewer: Gerrit User 131 <131@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0