Update patch set 1
Patch Set 1:

This fix is absolutely essential.  When Borislav modified the charts before they were merged, he made a breaking change to the input files that I did not notice when I did the review.
I do not know what is causing the problem you are seeing--the log output is from a rerunning of the bootstrap job, and the error is related to the fact that the job had been run before.   My original charts had configured the job for zero retries; Borislav said it did not work in his environment and commented out that configuration.

Patch-set: 1
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0