Update patch set 1
Patch Set 1:

Mike, this submission is not related to the errors you saw.  I think the error messages you quoted are due to the re-spin of the bootstrap container.  Its first life already added the plugins.  Then the reincarnations also try to upload again to the same Cloudify Manger.  Hence those "already exits" messages.   The root cause is the fist bootstrap failure, which could be due to what this submission fixes because in our experiments the old values would cause problems.

Patch-set: 1