Update patch set 5
Patch Set 5:

(2 comments)

Thanks for the review guys!
I was able to bring up Policy and the xDemoApp with the changes to launch script.  I also realized that I hardcoded the namespace and was missing the helm release name so I added 2 parameters.
They fail the first time you click to launch the app, but if you close them and re-launch they load just fine.

Patch-set: 5
Label: Code-Review=0