Update patch set 2
Patch Set 2:

I tried to squash/combine this and Jonathan's latest patch together with some of my changes and it ended up in another review.  Take a look, test it out and if it is good then we can abandon this and merge this one: https://gerrit.onap.org/r/#/c/45923/3

Sorry for that - good news is that it preserved the commit messages.  git/gerrit complexity :(

Patch-set: 2