Update patch set 2
Patch Set 2:

(1 comment)

Hello Mayank, regarding do-start.sh.   I believe it best to remove it so not to shadow the official do-start.sh from the policy/drools-pdp itself.   It was out of sync, and it is the case as I tested it that the official one works with both heat and oom installation.  I think shadowing the official copy should be last resort, to avoid maintenance of multiple copies.   Some cases may have to be done, I suppose that's the reason you introduced it in the PE side, but is not the current case for pdp-d.

Patch-set: 2