Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > Hello Mayank, regarding do-start.sh.   I believe it best to remove
 > it so not to shadow the official do-start.sh from the
 > policy/drools-pdp itself.   It was out of sync, and it is the case
 > as I tested it that the official one works with both heat and oom
 > installation.  I think shadowing the official copy should be last
 > resort, to avoid maintenance of multiple copies.   Some cases may
 > have to be done, I suppose that's the reason you introduced it in
 > the PE side, but is not the current case for pdp-d.

Hi Jorge,

If you remember, we had a discussion around this. 
In the do-start.sh file in the image, there are checks for the ports to be up for mariadb which is hard coded name and it will fail if service name of mariadb is changed in future.
Thats why I had overridden this file and removed the checks for port to be up as K8s starts the dependent pods only when initcontainer is done.

If you still want to keep the do-start.sh from image then I would suggest to pass an env variable with service name of mariadb (and other services) and use that in the scripts.
There are 2 do-start.sh scripts as of now in oom and both can be removed if you take care of these wait-for-port thing.

Thx,
Mahendra

Patch-set: 2