Update patch set 1
Patch Set 1:

This is the first pass to get helm charts in sync with the manifest.  We need to ensure that the helm charts are deploying the versions specified in the manifest.  If something is broken, then the project teams are expected to either fix the manifest or fix the supporting config/variables in the helm charts.

Patch-set: 1