Update patch set 1
Patch Set 1:

Borislav, I went with K8s 1.8 because that was the version the integration team and documentation are testing on.  

The other reason I selected k8s 1.8 is because as of k8s 1.8 a patch for a CVE made all injected volumes read only and this exposed some problems with some containers that were trying to modify volumes/configmaps injected by k8s.  I see that this was also introduced in 1.7.14 so I will include k8s 1.8 and 1.7 above those versions.

Any issues with requiring helm 2.8?

Patch-set: 1