Update patch set 1
Patch Set 1:

Hey Sylvain,

Your proposal for using monger is correct - modeling it after what appc does for mysql.

Unfortunately, the helm chart for mariadb-galera is not yet ready for general consumption. It is being worked on for the Casablanca timeframe.  The best thing to do right now is to copy the embedded mariadb chart for either SO or Policy. These are the best examples right now (not clamp as mentioned previously) on how you can configure mariadb with an application specific db schema and initial dataset. One example uses a job.yaml to do this and the other uses an init container. A standardized way of doing this will be introduced in Casablanca as we migrate over to the mariadb cluster.

Patch-set: 1
Reviewer: Gerrit User 127 <127@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0