Update patch set 3
Patch Set 3:

Mahendra, I'm very concerned with these additional capabilities added 1 day before release, just because of the cosmetics of the PAP_PDP_URL (which was following the original accepted one).   

We should have let through my review got as it was working, and put this on top, so it can be tested independently   

Again, the discrepancy on the PAP_PDP_URL to make it pretty, not sure if worth the risks at this point in time.   Again, this is better in a different patch or in Casablanca.

Not sure what you said about restarting the PAP, if configuration changes happen.  If that's the case, I don't want to restart the PAP if I grow one PDP-X, and trigger a chain reaction, that most likely won't work.   Not sure if you I may have misunderstood that, but I will be surprise if that brings the system clean, but again I may have misunderstood you.

Another issue, eyeballing the change in the PAP_PDP_URL will be different than the original one, so may need to modify the java code, to make it work.   Note that the PAP_PDP_URL should end with ../pdp/, and here will end up with the following: testpdp,alpha123,testpdp,alpha123, which may or may not work, but configuration wise is not correct.

I suggest pulling this into another review, and get mine through.  I'll test it independently today on top of mine, it it works will merge it, if not redone for Casablanca. 

I need guidance from you on the exact set of commands for the upgrade that you want to test.

Patch-set: 3