Update patch set 5
Patch Set 5: Code-Review-1

> > Testing your patch with this one: https://gerrit.onap.org/r/#/c/44705/8,
 > > I see that the key for mysql is Values.geoEnabled yet in 44705,
 > the
 > > key is changed to Values.config.geoEnabled.
 > > Is this going to work as expected?
 > 
 > mysql values is defining the key as Values.geoEnabled...

Agreed that is the value key for the mysql chart.  My comment was asking which chart is setting this to true.  I don't see any chart that references this common/mysql chart ever setting mysql.geoEnabled=true so I question whether it works as expected.

Assuming it is sdnc, then the sdnc charts values file needs to set that in the mysql: section where other mysql keys are overridden.

Patch-set: 5
Label: Code-Review=-1