Update patch set 2
Patch Set 2: Code-Review-1

This is a workaround that can be applied in the end users own values file and IMO doesn't need to be checked in or it will become "the solution" and other charts will have some precedence to apply the same band-aid. 

I would prefer a better solution that detects the true startup of the database and its schema.  I added a comment to the JIRA with more info and logs.

Patch-set: 2
Label: Code-Review=-1