Update patch set 2
Patch Set 2:

On second thought, in the interest of time I would be okay with increasing the default from 10s to something higher, just not 10 minutes.  In my envs, I haven't seen it take that long to load.  We've done 90 seconds in other places as a new default.  How about 3-5 minutes?

Patch-set: 2