Update patch set 2
Patch Set 2:

I agree on the "right" solution.  However...

Given that it's nearly impossible for the DB to complete initialization in 10 seconds, should we not at least bump the default liveness initial delay up to a value that has a chance to work?

Otherwise we're delivering OOM with out-of-the-box corrupted portal databases and a Portal UI container that cannot start, and then have to train people on how and why they have to manually override and recover from an unusable default setting.

Unless someone can implement the "right" solution by RC2, which is tomorrow.

Patch-set: 2