Update patch set 3
Patch Set 3:

Mahendra, I gave it a test with the last patch you added.   These are my observations:

1. The expanded configuration in the PAP has duplicated configuration credentials for the last PDP.   The software tolerates this, and there's no harm in functionality.

2. After modifying replica count per "helm upgrade" command discussed in comments to upgrade a PDP0X, the system became pretty messy.   PAP goes down, the 2 new PDP-Xs comes up first, after a while, PAP takes a bit to come back up, only 2 of the 4 PDP-Xs are detected in the PAP as part of the group.    The policies are in some PAPs but not in others, so going through the service gives inconsistent results.

Even if everything was to work fine, one major issue I find with this is forcing bringing down the PAP when growing a PDP-X instance when which is unnecessary, and causes service disruption, in UI, and other update functionality that requires group information.

I'm quitting for the day now.   Based on this, I think we should go to my first patch, with the change in the PAP_PDP_URL line to not skip the 2, which you were right on, when you pointed it out.  So if you agree if you could pull out the latest patch from the review, I would revert back to the original one tomorrow morning not skipping the 2.

Let me know how you guys want to move forward.

Patch-set: 3
Label: Code-Review=0