Update patch set 1
Patch Set 1:

> actually not this simple - I still need to account for a single
 > node deploy and make the replicaCount parameterized (1 for single
 > node), 3 for 3+ nodes, or ideally n for n

You can have multiple logstash instances running on a single node although the performance of this might not be much better than a single instance. We should look at making logstash a DaemonSet such that it will automatically scale with the number of nodes in the cluster to achieve the n:n ratio you're looking for.

Patch-set: 1