Update patch set 1
Patch Set 1:

No policy.. it is just the pattern I see.  At this point projects are moving from latest to fixed numbers but that is probably because their code bases have been locked down and the churn has stopped.  If there is still churn in the dcaegen2 images then yes we can use latest until the release is cut whenever that is.

Patch-set: 1