Update patch set 1
Patch Set 1:

The portal-db-config container needs to wait until after the portal-db initialization is fully completed before it is allowed to run; otherwise it will execute against an incomplete schema and leave the system in an undefined state.  The same can be said for any other pods that depend on the database to be fully populated.  Until we can enhance the readiness probe to check for the completion of all the schema init scripts, we must increase the readiness delay to allow for the full initialization time.

Patch-set: 1