Update patch set 2
Patch Set 2:

The cdt proxy service in the appc container is on port 9191, you need to make sure you have the changes to the startODL.sh file (also in this commit) in order to run that.

I need to make a change to the javascript in cdt so that it will connect to the correct kubernetes port instead of 9090, like it is now.

Patch-set: 2
Label: Code-Review=0