Update patch set 1
Patch Set 1: Code-Review-1

I can't accept this as a solution, as it completely bypasses the usefulness of having a readiness probe.  If it takes APP-C 15 minutes to start, then it's not "ready" to serve requests until that's done.  

When an APP-C pod fails, and a new one is started to replace it, the new one could immediately begin receiving requests that it isn't ready to serve.

Patch-set: 1