Update patch set 1
Patch Set 1:

The existing port probe defeats the purpose of having a readiness probe, because it is indicating that APPC is ready when it really isn't. This change is actually making the readiness probe work as it was designed to so that the appc pod is not flagged as ready until it is actually ready to handle requests.

Patch-set: 1
Label: Code-Review=0