Update patch set 1
Patch Set 1: -Code-Review

Was the issue that the readiness probe was marking APP-C ready too soon?  If so, that's likely happening because ODL is started twice in startODL.sh, and it passed readiness sometime during the initial up-time while features are being installed.  If that's the case, then you're right, it's not a very good readiness test and we should devise a better one.

With this solution, the health_check.sh file is created ( and available for the readiness check to run ) before the second time the karaf executable is invoked.  So there's definitely still an opportunity to advertise readiness before it's time.

Could we instead have a readiness test that hits an API which is only available after karaf is fully up the second time it's started?

I expect that SDN-C suffers from a similar problem, but it may not be as obvious since it installs fewer bundles and starts quicker.

Patch-set: 1