Update patch set 3
Patch Set 3:

Mayank, 
do you have a solution here? We are forced to go this route for CDT due to constraints with ODL discovered late in the cycle. We need our proxy to allow CDT to connect to APPC. OOM makes it difficult since ports can change. If you know of a way to make it work reliable and can validate it yourself, let us know. We need to get this committed today.

Patch-set: 3