Update patch set 1
Patch Set 1:

portal-app is an internal kubernetes image. The port however is an exposed port for external use. That is the reason we are just trying to make the port configurable at this point. For Casablanca, We will try to also make the image names configurable

Patch-set: 1
Reviewer: Gerrit User 679 <679@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0