Update patch set 2
Patch Set 2:

Not sure Borislav, I'll follow your advice on this, I may have misunderstand something.   I see in the policy/charts/drools/templates/statefulset.yaml, the following:

          readinessProbe:
            tcpSocket:
              port: {{ .Values.service.externalPort }}
            initialDelaySeconds: {{ .Values.readiness.initialDelaySeconds }}
            periodSeconds: {{ .Values.readiness.periodSeconds }}

The problem that I see in Marco's lab, is that part of initialization of the container is trying to download a zip file.   The network in windriver seems extremely slow.   What I see is the container being restarted during this download process (without completing).  This is before the actual application process is started (which in turn will open the port that will be monitored in tcpSocket section above).  Based on the previous helm section I was assuming readiness.

Let me know what you think is best.   Thank you.

Patch-set: 2
Label: Code-Review=0