Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > Yes, finally.  :)

Since the authorization file is injected as a configMap, it is ready-only by default and can't be changed.  I think the way forward is to change whatever is trying to modify this file or we apply some k8s techniques to inject and copy the file so that it isn't read only.

Please raise a separate bug if there isn't one already for the init_robot issue.

Patch-set: 1
Label: Code-Review=0