Update patch set 1
Patch Set 1: Code-Review-1

It is supposed to proxy through nexus3.onap.org to dockerhub so the original values are indeed correct.

Is there an error during image pull?  what was the driver for this change?  If there is an error in image pull, then we should open an LF ticket to investigate.

Patch-set: 1
Reviewer: Gerrit User 131 <131@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1