Update patch set 4
Patch Set 4:

Hi Jasmine, just uploaded another patch set, I think this should be the last. Regarding the naming conventions, i agree they are inconsistent but things are functional and i dont see the harm in leaving as-is. The same for the volumes, core functionality is working, i dont think we need to investigate further on it.

Patch-set: 4
Label: Code-Review=0