Update patch set 1
Patch Set 1:

Hello Durgpal, the helm changes are very clever, but I have a major concern with the runtime behavior.  Let me know if I have misunderstood something about it.

So, if I understand correctly the changes, when a new pdp is grown, it will cause a reconfiguration of the pap, taking a healthy pap down with it (not sure if there may be other components).   This in turn, will trigger synchronization between policy components and will introduce at least temporary instability.  If my assumption is correct, the scale out operation would have the opposite effect, it will reduce availability of the system.   This is in contrast with the current way of operation, where if a pdp is scaled out, it will not take down the pap, or any other healthy components.

I agree that this existing helm code in the pap.conf is sub-optimal, but I am afraid that  the only reasonable solution is to modify the whitelisting functionality in the java code instead, after that, we can remove this piece of configuration from the helm charts.

Let me know your thoughts.

Patch-set: 1