Added external-key and resource-name custom fields
During the self serve flow, the resources can be uniquely
identify using a external-key, allowing the DG to retrieve
the data from SDNC DB very easily for unassign purposes.
The resource-name has been added in order to provide more
information for the netbox admin, allowing him to have a
clear idea what purpose the ip is serving; given the
resource-name passed on by the designed is meaningfull.

Change-Id: I33bb74b9bca8009e70faede8001c34a451ad205f
Issue-ID: CCSDK-366
Signed-off-by: Alexis de Talhouët <adetalhouet89@gmail.com>