Update patch set 2
Patch Set 2:

> > Uploaded patch set 3.
 > 
 > This is to update the nodePort. Notice this changes the range via
 > the nodePortPrefix in values.yaml to 304 as 302 is exhausted.
 > Together with this there is an update to service.yaml which I am
 > not entirely sure about.  Would be grateful for any comments.

This is a bit problematic, because if global.nodePortPrefix will be overwritten (using command line for example) the chart nodePortPrefix will remain.
Currently there is a discussion in oom team regarding what to do with nodePorts. Hopefully it will be resolved soon. I do not see port 77 used except in the wiki. Maybe it would be better to ask in the distribution list oof team whether they need this port and if not, use port 30277 (and update it in the wiki).

Patch-set: 2
Label: Code-Review=0