Update patch set 1
Patch Set 1: Code-Review-1

Yes, I think we should not merge this code.  It is very elegant but the side effect of taking the PAP down and the subsequent instability introduced when scaling out a PDP-X, I believe it to be severe enough to discourage the merge of this code.

I think the best solution to the issue that OOM-850 reports, is to do equivalent functionality in the policy side, PAP java code to avoid system disruption.   I created a task on the policy side to track it:

https://jira.onap.org/browse/POLICY-1087

After that is done, the conservative bulk preprovisioning in the oom side in kubernetes/policy/resources/config/pe/pap.conf, could be removed.

Patch-set: 1
Label: Code-Review=-1