Update patch set 3
Patch Set 3:

Hello all,

For this task:-

Now, We can see (Many new sub charts added for SO) Which contains their specific defined Resource limit,  

So Do we want to introduce segregation of flavor in same?
We can use the given values as Small flavor and put the same as by default in order to keep it as it is and can give Large Flavor with high config liberty, Please let me know your views?

BR/
VC

Patch-set: 3
Label: Code-Review=0