Update patch set 15
Patch Set 15:

looking at this now.  I had AAI running without this change and it does not upgrade:
3       	Fri Sep  7 15:10:43 2018	SUPERSEDED	aai-2.0.0	Upgrade "aai" failed: Job.batch "aai-aai-traversal-update-query-data" is invalid: spec.template: Invalid value: core.PodTemplateSpec{ObjectMeta:v1.ObjectMeta{Name:"aai-traversal", GenerateName:"", Namespace:"", SelfLink:"", UID:"", ResourceVersion:"", Generation:0, CreationTimestamp:v1.Time{Time:time.Time{wall:0x0, ext:0, loc:(*time.Location)(nil)}}, DeletionTimestamp:(*v1.Time)(nil), DeletionGracePeriodSeconds:(*int64)(nil), Labels:map[string]string{"controller-uid":"37183b29-b207-11e8-98ca-...

Is it okay with you all, that to deploy this, everyone will have to undeploy the AAI helm chart and re-deploy it?  Persisted data should still be intact...

There are certain things helm cannot automatically upgrade and something in the Jobs introduced by this patch are one of those things....

Patch-set: 15