Update patch set 15
Patch Set 15:

by before, i assume you meant before the latest patch that addressed review comments..then the answer is i didn't test that.  when i first deployed this it was a fresh install aai with these changes.  what i showed previously was an aai that did not have these changes and the upgrade was introducing them.

Patch-set: 15