Update patch set 8
Patch Set 8:

@Pramod can I ask that you please split this patch into 2? 
You can create a new patch that is strictly for the mariadb-galera chart (which we can merge immediately). And then amend this patch to remove mariadb-galera changes and leave only the appc changes. There is another chart that is now dependent on mariadb-galera and I would like to get that in asap.

Patch-set: 8