Add helm plugins to (un)deploy onap
In the Casablanca release, ONAP has become very large.
Every Helm chart contains some amount of external configuration.
And unfortunately, there is a limit to the amount of configuration that
can exist in a Helm chart. When ONAP is installed as a unbrella chart
containing many subcharts (each with configuration), the
amount of configuration that resides in configmaps in K8s
causes the installation to fail. To work around this issue, it was
necessary to introiduce a Helm plugin that will install/upgrade onap
by deploying the parent chart and each subchart within its own
Helm "release". This is the initial version of the plugins, which
has been verified to work and unblock integration in the deployment
of all of ONAP. Improvements will be made post M4 to further harden
and enhance the plugins. Specifically, the plugins are not currently
resilent to network issues which can cause some of the subcharts to
fail to deploy. To work around this, the plugins should be run from
within the same network as the k8s cluster you are deploying to.

Change-Id: I7c586f1c75f7249d2d7c38b4088129162c96401d
Issue-ID: OOM-1344
Signed-off-by: Mike Elliott <mike.elliott@amdocs.com>