Update patch set 1
Patch Set 1:

We need to be updating the default image tags within the charts themselves.  Alexis, please drop the cert injection and see if it is present within the latest version of the container as Brian says...Then we can merge the updated versions.

The script that subs the tags is great but IMO it is an integration specific convenience tool and I don't think we can expect the whole community to use it.

In previous releases it was a pain but a few of us manually synched the charts with the manifest.  In hindsight, we should have flipped all charts to latest or whatever 1.3. staging-latest at the start of the Casblanca dev cycle...It is actually kind of scary that people including myself are doing code reviews and tests using Beijing images.

Patch-set: 1