Update patch set 2
Patch Set 2:

> OPTFRA-299 is more about hardening I guess (and seems to be a
 > closed issue). Just curious, why are we adding AAF here? Can we
 > create a different issue / tasks with better details in it?

aaf-service is needed for AAF RBAC and authentication,  aaf-sms is required for Secret Management Service. OPTFRA-299 was closed sooner I guess. Anyway I have created another task for this under the epic for hardening as I couldn't find a better suited Epic in the list. Please suggest if this should be associated with another Epic.

Patch-set: 2