misc. changes to pass robot aaf healthchecks
It tackles several problems found in the helm charts
that prevented instantiation of aaf-cm, aaf-fs, aaf-gui,
aaf-locate, aaf-oauth, and aaf-service pods.

There is duplication of org.osaaf.aaf.* files that
match the authz repo expected locations in java code.
The definite location is left up to the AAF SMEs for
future reviews.  Here the files are duplicated to make
sure aaf components start.

Change-Id: I78ae8e18518fe6dad856bd5af01bc8ff4fa71644
Issue-ID: AAF-538
Signed-off-by: Jorge Hernandez <jh1730@att.com>