Update patch set 2
Patch Set 2:

> new so-monitering container. there is a confd configmap that does
 > not start with dev-so created so it doesnt auto-delete and helm del
 > dev-so the container sticks in terminating. Also doesn't come up on
 > helm deploy stuck in init phase
Hi Brian, This SO monitoring pod has been created by Ericsson team and their configmap was missing an application configmap and override file. With the above changes at SO env, it is able to go past init phase. However the application doesn't start fully, fails for health checks and I don't see any application logs to identify the issue with the application. I am working with Rob to find the root cause of this issue. 
We need this CatalogDBAdapater change in urgently as this is blocking So testing. Can you please approve this and I will ensure that this so monitoring issue is addressed shortly.

Patch-set: 2
Reviewer: Gerrit User 4277 <4277@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0