Update patch set 17
Patch Set 17: Code-Review-1

The cassandra-job seems to be running /runcql.sh which times out in our cluster when we are trying to bring up music. 
The default cqlsh timeout is 10. Maybe we should look at using a larger timeout in the runcql.sh script so the script does not timeout in between.
Any errors or timeouts that happen in between become unrecoverable a the admin password gets changed in the second step. This means that we have to bring up the entire music and sub charts again.

Patch-set: 17
Label: Code-Review=-1