Update patch set 7
Patch Set 7:

There were comments made about addressing some deficiencies with this chart in the Dublin timeframe. If that is the go forward plan, can a please ask that jira task for that work be created and linked to this patch as a comment? A task in OOM-1194 would do nicely.

Patch-set: 7
Label: Code-Review=0