Update patch set 1
Patch Set 1:

The issue of so-monitoring was traced to be a mismatch between the application port and the container port. So I changed all the override files to match application port with the container port in the values file. so monitoring pod seems to work correct now. Please review.

Patch-set: 1
Reviewer: Gerrit User 4277 <4277@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0