Provision clients for PNF topics
In addition to adding the pub and sub client Role to the PNF topics
for use by the PRH, we fixed the readinessProbe to wait for the
message-router pod.  This should avoid attempts to provision before MR
is ready to accept.

Also, properly delivered the dmaapbc.properties file which was lost from
previous commits!

Change-Id: I8016bc83a9c3ba75de22732e475c5f1896ea4ae4
Signed-off-by: dglFromAtt <dgl@research.att.com>
Issue-ID: DMAAP-548