Update patch set 9
Patch Set 9: Code-Review-1

(1 comment)

That oom variable substitution is what will break spring – it looks like a the nodeport was not tested and ctxBuilderExternapPort is not named properly or it is missing or mis-spelled in values.yaml in the sub chart 
check the oom patch for sdnc

"{{ .Values.config.ctxBuilderExternalPort }}"

then check all the already merged code that calls the sdnc-ctx – that they have been retrofitted to call the new sdnc container properly.


For the sdnc (target) patch there is no external node port 302xx – as expected because we use the pomba k8s service name inside the pod
https://gerrit.onap.org/r/#/c/55451/9/kubernetes/pomba/charts/pomba-sdncctxbuilder/templates/deployment.yaml

Check that the variable in the calling service is using 9530 and the dns name in the calling chart that references sdnc-ctx-builder – defined below
https://gerrit.onap.org/r/#/c/55451/1/kubernetes/pomba/charts/pomba-sdncctxbuilder/values.yaml 


From: J.Ram  
Sent: Thursday, October 18, 2018 10:17 PM
Cc: *
Subject: “BugFix: missing variable in SDNC OOM” -- Fixed

1126

The SDNC CtxBuilder is having logging issues. 
I seem to have a problem in the CTX Agg, with the below error: 

2018-10-19 01:47:39,352|||pool-2-thread-2||||ERROR||||||| Error getting result from sdnc context builder.  Reason - For input string: "{{ .Values.config.ctxBuilderExternalPort }}"
2018-10-19 01:47:39,352|||pool-2-thread-2||||INFO||||||| Error returned from one of the Context builders, no event will be published.

1.	The defect above shows that the logging is not configured properly for sdnc ctx builder as it is deployed with aai-ctx-builder logging.  And there is 0 logs to show any information from the error above 
2.	The actual error, I think is a separate bug.  Designers please advise ( from point 1, I cannot say what the real issue is)

Patch-set: 9