Update patch set 4
Patch Set 4:

Mike, you are right that it should only be enabled false for dev environment when you don't want aaf locally as it could be resource intensive to run that many containers locally. What I did was added a aaf.enabled flag local to chart aai-traversal and the global values file should override it.

Patch-set: 4