Update patch set 1
Patch Set 1: Code-Review+1

Yes, the workaround was originally for saturation of logstash - after looking into it post the Daemonset change - it turned out the CPU saturation is only when AAI is on the same node - in this case the secondary workaround is to limit the vCPU to 3 until I fix the root cause.  We will revisit the daemonset when we bring in a file instead of beats plugin to get the logs directly from docker.  For now yes we should revert this because the daemonset is not being used correctly anyways for traffic - I like the labels idea better

Patch-set: 1
Label: Code-Review=+1