Update patch set 2
Patch Set 2:

> (3 comments)
 > 
 > Some minor comments

Thank you for the comments Eric.

I have removed the trailing white space and corrected the spelling, but I did not need to interact with the dmaap-bc so did not. There is a nodeport relating to it though (30241) but it doesn't seem necessary at this point in time.

Patch-set: 2
Reviewer: Gerrit User 4327 <4327@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0