Update patch set 1
Patch Set 1:

> There was - I don't think you got added. Sorry about that Mike. We
 > can always submit another review to make any fixes you suggest.

No problem.. My comment on this one is optional really.  It will work either way but just wanted to point out the policyScope field in the json is deprecated and the code doesn't do anything with it.  We now just add the scope to the policyName.. no biggie really, it won't break anything.

Patch-set: 1