Update patch set 11
Patch Set 11:

Looks like this commit introduces a bug after all (missed that in reviews).
It uses port 30263 in multicloud/charts/multicloud-prometheus/charts/prometheus-grafana/values.yaml which is also used in sdc-dcae-fe chart (and is allicated in node-port wiki page to sdc)
This causes sdc to fail to deploy.

Patch-set: 11