Update patch set 1
Patch Set 1:

Teams probably weren't expecting to have to individually submit the image changes in OOM directly since this sync method was the process used for the original Casablanca release.  I'm inclined to suggest that let's keep the same process for Casablanca MR, and formally announce and change the process for Dublin.

Patch-set: 1