Update patch set 2
Patch Set 2:

> > Hi Borislav,
 > >
 > > I have made some amendments in a new patchset to add some missing
 > > config, but I have not yet moved the common certificates up into
 > > the parent chart.
 > >
 > > Since other securitized microservices (e.g. aai-resources) are
 > all
 > > following the same structure (with certificates in the child
 > > charts), may I suggest that you merge this change first, and we
 > > will follow up with a subsequent changelist where we will update
 > > all securitized microservices to source their common certificates
 > > from the parent chart?
 > 
 > Sorry, no. I prefer you first make order with all those p.12 files
 > and then add support in other services in a proper way.
 > Also, I would like to issue a discussion regarding all this fproxy
 > and rproxy features. I feel they clutter aai charts and need to
 > make sure we are doing the right thing. Mike, for your
 > consideration.

Hi Borislav,

I appreciate your concern about the duplication of the p12s, but as things stand now, Gizmo and Champ will not start up when deployed with OOM, because they need these config changes in place.

Restructuring the charts to reduce duplication is a valid suggestion, but it is not a trivial task and we would have to take this into our sprint as a new story to be completed in order of priority.

So in the mean time, while acknowledging that this is not the perfect solution, it would at least allow Gizmo and Champ to function again.

Patch-set: 2
Label: Code-Review=0