Update patch set 1
Patch Set 1:

> Aligns pretty close to my last run http://jenkins.onap.info/job/oom-cd-master2-aws/111/consoleFull
 > 
 > your pods and hc are good
 > +1

Hi Mike,

We are still seeing issues with the maria-db container not behaving correctly. 
We expect the db, user etc to be created on startup but it seems to fail and restart  once for some reason. This causes the dmaap-dr-prov to fail to initialise the DB tables.

Seems to be environment specific problem also as we cannot replicate it here.

Would you have any guidance on this or should we just try to move away from having our own db chart?

Patch-set: 1
Reviewer: Gerrit User 4405 <4405@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0