Update patch set 2
Patch Set 2:

The image name problem is due to an incompatibility between the script that integration uses to change all of the image versions in the Helm charts and the way the inventory (and service change handler) Helm charts are written.   We are looking at changes to charts so that the integration script will not generate bad image tags.

Patch-set: 2
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0